requirements.txt
openai
python-dotenv
pdfplumber 
pymupdf
pandas
streamlit